LayoutTests/imported/w3c/web-platform-tests/css/css-logical/animation-004-expected.txt


Harness Error (TIMEOUT), message = null

FAIL Logical properties can be transitioned assert_equals: expected "50px" but got "100px"
FAIL Logical properties in transitions respect the writing-mode assert_equals: expected "50px" but got "100px"
FAIL Logical properties in transitions respect the direction assert_equals: expected "50px" but got "100px"
PASS Declaration order is respected within declaration blocks
PASS Logical properties are able to override physical properties in declaration blocks
PASS Declaration order is respected amongst logical properties within declaration blocks
PASS Physical properties and logical properties can be mixed
PASS Declaration order is respected on each keyframe individually
FAIL Transitions update when the writing-mode is changed assert_equals: expected "50px" but got "100px"
TIMEOUT Filling transitions update when the writing-mode is changed Test timed out
FAIL The number of interpolating properties can be increased when the writing-mode is changed assert_equals: expected "50px" but got "100px"
FAIL The number of interpolating properties can be decreased when the writing-mode is changed assert_equals: expected "150px" but got "200px"
FAIL Transitions update when the writing-mode is changed through a CSS variable assert_equals: expected "50px" but got "100px"
FAIL Transitions update when the direction is changed assert_equals: expected "50px" but got "100px"
PASS Transitions from logical to physical update when the direction is changed
FAIL Transitions from physical to logical update when the direction is changed assert_equals: expected "150px" but got "100px"
